NotificationCompat ; check URLs ; skip 404 files
Switched to NotificationCompat so that earlier devices are supported.
Note that MediaPlayer prior to 3.1 doesn't support HTTPS urls, so not
all features will be available in all versions.

URLs are checked for basic formatting when the M3u playlist is parsed.
Invalid URLs are ignored. Similarly, if a URL returns a 404, the track
is skipped.